rrr: convert rpc/utils to module (rrr.utils)
utils.cpp becomes the rrr.utils module interface unit; utils.hpp
deleted. AddrInfo RAII wrapper + set_nonblocking / find_open_port /
get_host_name free functions. Uses `import rrr.logging;` for the
Log_* calls.

Includers (2) updated:
  - src/rrr/rpc/server.hpp, src/rrr/rrr.hpp.

Measurement (rrr + rpcbench, -j32, clean):
  +epoll_wrapper:  39.01 s, 167.1 MB PCM
  +utils:          38.07 s, 174.9 MB PCM, librrr.a 9.62 MB

Wallclock came in slightly under the prior step (noise). 9 modules,
2.26× baseline. 3× threshold (~50 s) still has margin.
